refactor: move dir scanning
